+++ lib/duckdb/converter.rb
@@ -8,6 +8,11 @@ module DuckDB
   QueryProgress = Struct.new(:percentage, :rows_processed, :total_rows_to_process)
 
   module Converter # :nodoc: all
+    RANGE_INT8 = -128..127
+    RANGE_INT16 = -32_768..32_767
+    RANGE_INT32 = -2_147_483_648..2_147_483_647
+    RANGE_INT64 = -9_223_372_036_854_775_808..9_223_372_036_854_775_807
+
     HALF_HUGEINT_BIT = 64
     HALF_HUGEINT = 1 << HALF_HUGEINT_BIT
     LOWER_HUGEINT_MASK = HALF_HUGEINT - 1